+++ api/Endpoints/Limeyfy/Invoices/Add/Models.cs
@@ -10,9 +10,9 @@ public class Request
     
     public string CompanyId { get; set; }
     
-    public int OrganizationId { get; set; }
+    public long OrganizationId { get; set; }
     
-    public int BankAccount { get; set; }
+    public long BankAccount { get; set; }
     
     public int InvoiceNumber { get; set; }
